Implements barring of users
Should a student login once on one computer, then again
on the other, on assigning session values for cate access
we invalid the first user.
This commit implements barring of the first user, and so
the first user will be redirected to the login page with an
alert specifying that another user has logged in.